api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = "API Gateway for E-Commerce Microservices Platform"
authors = ["Minh Duc <mduc24@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
httpx = "^0.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
